CSV output and deleted unneeded files
The script can now save the name, date, and time of each person it recognizes. Then orders it alphabetically and deletes duplicate identifications. It outputs the data to FaceReqData.csv which is located in the same path as the script.

headshots.py, headshot_picam.py, and sent_test_email.py were deleted.